[mka]: Fix re-establishment by reset MI (#72)
Key server may be removed due to the ingress packets delay. In this situation, the endpoint of key server may not be aware of this participant who has removed the key server from the peer list. Because the egress traffic is normal, the key server will not remove this participant from the peer list of key server. So in the next MKA message, the key server will not dispatch a new SAK to this participant.
And this participant can not be aware that is a new round of communication so it will not update its mi at re-adding the key server to its peer list.
So we need to update mi to avoid the failure of the re-establishment MKA session.

Signed-off-by: Ze Gan <ganze718@gmail.com>